requirements.txt
pandas==2.1.0
openpyxl==3.1.2
openai==1.3.0
streamlit==1.28.0
python-dotenv==1.0.0 